Fix parameter passing (#31)
Pass parameters to script without interpretation

'$*' will run parameters through globbing, so the string '.*' will
become a list of filenames. This means
  rabbitmqctl set_permissions -p / admin '.*' '.*' '.*'
could be turned into
  $PREFIX/rabbitmqctl set_permissions -p / admin . .. .mkdir . .. .mkdir . .. .mkdir

"$*" would pass all parameters as a single string, which also doesn't
work. Instead use $@, which passes the parameters as a list, and
specifically "$@", which passes the parameters as a list without
applying any expansions.